pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "google-generativeai>=0.3.0"
]
